Move bearing into maths
